modify(ResponseLine): HttpStatus를 변경이 가능하도록 수정
- ResponseLine의 형식으로 가져오는지 테스트 작성
